moved all files within the layout folder into the components folder. updated all file paths
