+++ client/src/components/LoginPage/Login.scss
@@ -1,5 +1,5 @@
 // importing application wide styles
-@import '../layout/Style.scss';
+@import '../Style.scss';
 
 //login screen
 .form-div{
