+++ client/src/components/Profile.js
@@ -2,9 +2,9 @@ import React, { Component } from "react";
 import { Container } from 'react-bootstrap'
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { logoutUser } from "../../actions/authActions";
-import { searchProjects } from '../../actions/index';
-import ProjectProfileList from '../containers/ProjectProfileList';
+import { logoutUser } from "../actions/authActions";
+import { searchProjects } from '../actions/index';
+import ProjectProfileList from './containers/ProjectProfileList';
 
 
 class Profile extends Component {
