+++ client/src/components/Style.scss
