+++ client/src/components/redux/AddProject.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import {Container} from 'react-bootstrap'
 import CreateProject from '../containers/CreateProject';
 
-import '../layout/Style.scss';
+import '../Style.scss';
 
 class AddProject extends Component {
   render() {
