+++ client/src/components/redux/ViewProfileProject.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import '../layout/Style.scss';
+import '../Style.scss';
 import ViewAProfileProject from '../containers/ViewAProfileProject';
