+++ client/src/components/redux/ViewProject.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Container } from 'react-bootstrap'
 import { connect } from 'react-redux';
-import '../layout/Style.scss';
+import '../Style.scss';
 import ViewAProject from '../containers/ViewAProject';
